WIP: Update the provided grammar
This starts the process of updating the sample grammar that was
provided to include all the features of the unnamed language.
At present, it is not functional because of a non-LL(*)  compliant
expression rule.

It should also be noted that operator precedence is not yet enforced.
Solving the latter issue should also resolve the former.