Add a TypeNode class with a temporary string member
The TypeNode class will need to be refactored once we have the necessary
Type infrastructure in place. Since I don't yet know how to do this,
I've just added a string field that reports what was matched so we can
move forward with modelling the function infrastructure.

A TypeNode field is added to the FunctionDecl node.