Add tests for the Environment infrastructure
Adds simple unit tests using assertions in the main method
of semantic.Environment. Replaces the Stack with a Deque.

Adds a unit_test target to the build which will be used to
run all such unit tests that are added in the future, and
updates the README to include this information.