Removes gotos and returns which immediately follow others
This optimization adds a boolean flag to the code generator which
removes unreachable return statements and GOTO's. This prevents
the JVM from causing runtime errors in some cases where there is
always a return statement, but where there is also a goto that will
never run, but if it did then it would bypass all returns.

This is not an ideal solution because it isn't very clean and is
error-prone, but doing proper dataflow analysis is out of the scope
of the project.